Fix the action button's text when cards are selected
This used to work because clicking on a card would not redirect the
user to a different pane. Since it's no longer true, we should rather
specify the action button's text anytime the card is drawn so
navigating through the different tabs of the application consistently
display the card's state.